fix: add search index prop
